requirements.txt
google-generativeai>=0.8.0
streamlit>=1.44.0
plotly>=6.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
